+++ slasher/detection/attestations/slashing_bench_test.go
@@ -156,7 +156,7 @@ func BenchmarkCheckAttestations(b *testing.B) {
 		ia1.Data.Slot = (i + 1) * params.BeaconConfig().SlotsPerEpoch
 		root := []byte(strconv.Itoa(int(i)))
 		ia1.Data.BeaconBlockRoot = append(root, ia1.Data.BeaconBlockRoot[len(root):]...)
-		if _, err := detector.IsSlashableAttestation(context, ia1); err != nil {
+		if _, err := detector.DetectAttestationForSlashings(context, ia1); err != nil {
 			b.Errorf("Could not call RPC method: %v", err)
 		}
 	}
