+++ slasher/rpc/BUILD.bazel
@@ -8,11 +8,12 @@ go_library(
     deps = [
         "//beacon-chain/core/helpers:go_default_library",
         "//proto/slashing:go_default_library",
-        "//shared/hashutil:go_default_library",
         "//slasher/db:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
+        "@org_golang_google_grpc//codes:go_default_library",
+        "@org_golang_google_grpc//status:go_default_library",
     ],
 )
 
@@ -22,7 +23,6 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//proto/slashing:go_default_library",
-        "//shared/params:go_default_library",
         "//slasher/db:go_default_library",
         "@com_github_gogo_protobuf//proto:go_default_library",
         "@com_github_prysmaticlabs_ethereumapis//eth/v1alpha1:go_default_library",
