+++ slasher/rpc/server.go
@@ -2,10 +2,6 @@ package rpc
 
 import (
 	"context"
-	"fmt"
-	"sync"
-
-	"github.com/prysmaticlabs/prysm/shared/hashutil"
 
 	"github.com/gogo/protobuf/proto"
 	"github.com/pkg/errors"
@@ -13,6 +9,8 @@ import (
 	"github.com/prysmaticlabs/prysm/beacon-chain/core/helpers"
 	slashpb "github.com/prysmaticlabs/prysm/proto/slashing"
 	"github.com/prysmaticlabs/prysm/slasher/db"
+	"google.golang.org/grpc/codes"
+	"google.golang.org/grpc/status"
 )
 
 // Server defines a server implementation of the gRPC Slasher service,
@@ -25,65 +23,7 @@ type Server struct {
 // IsSlashableAttestation returns an attester slashing if the attestation submitted
 // is a slashable vote.
 func (ss *Server) IsSlashableAttestation(ctx context.Context, req *ethpb.IndexedAttestation) (*slashpb.AttesterSlashingResponse, error) {
-	//TODO(#3133): add signature validation
-	if req.Data == nil {
-		return nil, fmt.Errorf("cant hash nil data in indexed attestation")
-	}
-	if err := ss.SlasherDB.SaveIndexedAttestation(req); err != nil {
-		return nil, err
-	}
-	indices := req.AttestingIndices
-	root, err := hashutil.HashProto(req.Data)
-	if err != nil {
-		return nil, err
-	}
-	attSlashingResp := &slashpb.AttesterSlashingResponse{}
-	attSlashings := make(chan []*ethpb.AttesterSlashing, len(indices))
-	errorChans := make(chan error, len(indices))
-	var wg sync.WaitGroup
-	lastIdx := int64(-1)
-	for _, idx := range indices {
-		if int64(idx) <= lastIdx {
-			return nil, fmt.Errorf("indexed attestation contains repeated or non sorted ids")
-		}
-		wg.Add(1)
-		go func(idx uint64, root [32]byte, req *ethpb.IndexedAttestation) {
-			atts, err := ss.SlasherDB.DoubleVotes(idx, root[:], req)
-			if err != nil {
-				errorChans <- err
-				wg.Done()
-				return
-			}
-			if atts != nil && len(atts) > 0 {
-				attSlashings <- atts
-			}
-			atts, err = ss.DetectSurroundVotes(ctx, idx, req)
-			if err != nil {
-				errorChans <- err
-				wg.Done()
-				return
-			}
-			if atts != nil && len(atts) > 0 {
-				attSlashings <- atts
-			}
-			wg.Done()
-			return
-		}(idx, root, req)
-	}
-	wg.Wait()
-	close(errorChans)
-	close(attSlashings)
-	for e := range errorChans {
-		if err != nil {
-			err = fmt.Errorf(err.Error() + " : " + e.Error())
-			continue
-		}
-		err = e
-	}
-	for atts := range attSlashings {
-		attSlashingResp.AttesterSlashing = append(attSlashingResp.AttesterSlashing, atts...)
-	}
-	return attSlashingResp, err
+	return nil, status.Errorf(codes.Unimplemented, "method IsSlashableAttestation not implemented")
 }
 
 // IsSlashableBlock returns a proposer slashing if the block header submitted is
@@ -138,53 +78,5 @@ func (ss *Server) AttesterSlashings(ctx context.Context, st *slashpb.SlashingSta
 // DetectSurroundVotes is a method used to return the attestation that were detected
 // by min max surround detection method.
 func (ss *Server) DetectSurroundVotes(ctx context.Context, validatorIdx uint64, req *ethpb.IndexedAttestation) ([]*ethpb.AttesterSlashing, error) {
-	spanMap, err := ss.SlasherDB.ValidatorSpansMap(validatorIdx)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to get validator spans map")
-	}
-	minTargetEpoch, spanMap, err := ss.DetectAndUpdateMinEpochSpan(ctx, req.Data.Source.Epoch, req.Data.Target.Epoch, validatorIdx, spanMap)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to update min spans")
-	}
-	maxTargetEpoch, spanMap, err := ss.DetectAndUpdateMaxEpochSpan(ctx, req.Data.Source.Epoch, req.Data.Target.Epoch, validatorIdx, spanMap)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to update max spans")
-	}
-	if err := ss.SlasherDB.SaveValidatorSpansMap(validatorIdx, spanMap); err != nil {
-		return nil, errors.Wrap(err, "failed to save validator spans map")
-	}
-
-	var as []*ethpb.AttesterSlashing
-	if minTargetEpoch > 0 {
-		attestations, err := ss.SlasherDB.IdxAttsForTargetFromID(minTargetEpoch, validatorIdx)
-		if err != nil {
-			return nil, err
-		}
-		for _, ia := range attestations {
-			if ia.Data == nil {
-				continue
-			}
-			if ia.Data.Source.Epoch > req.Data.Source.Epoch && ia.Data.Target.Epoch < req.Data.Target.Epoch {
-				as = append(as, &ethpb.AttesterSlashing{
-					Attestation_1: req,
-					Attestation_2: ia,
-				})
-			}
-		}
-	}
-	if maxTargetEpoch > 0 {
-		attestations, err := ss.SlasherDB.IdxAttsForTargetFromID(maxTargetEpoch, validatorIdx)
-		if err != nil {
-			return nil, err
-		}
-		for _, ia := range attestations {
-			if ia.Data.Source.Epoch < req.Data.Source.Epoch && ia.Data.Target.Epoch > req.Data.Target.Epoch {
-				as = append(as, &ethpb.AttesterSlashing{
-					Attestation_1: req,
-					Attestation_2: ia,
-				})
-			}
-		}
-	}
-	return as, nil
+	return nil, status.Errorf(codes.Unimplemented, "method IsSlashableAttestation not implemented")
 }
